#6 Clean Up UI
