+++ WebApplicationMVC/Controllers/HomeController.cs
@@ -7,55 +7,15 @@ namespace WebApplicationMVC.Controllers
     public class HomeController : Controller
     {
         private readonly ILogger<HomeController> _logger;
-        private static readonly IList<CommentModel> _comments;
 
-        static HomeController()
+        public HomeController(ILogger<HomeController> logger)
         {
-            //_logger = logger;
-
-            _comments = new List<CommentModel>
-            {
-                new CommentModel
-                {
-                    Id = 1,
-                    Author = "Daniel Lo Nigro",
-                    Text = "Hello ReactJS.NET World!"
-                },
-                new CommentModel
-                {
-                    Id = 2,
-                    Author = "Pete Hunt",
-                    Text = "This is one comment"
-                },
-                new CommentModel
-                {
-                    Id = 3,
-                    Author = "Jordan Walke",
-                    Text = "This is *another* comment"
-                },
-            };
-        }
-
-        [Route("comments/new")]
-        [HttpPost]
-        public ActionResult AddComment(CommentModel comment)
-        {
-            // Create a fake ID for this comment
-            comment.Id = _comments.Count + 1;
-            _comments.Add(comment);
-            return Content("Success :)");
-        }
-
-        [Route("comments")]
-        [ResponseCache(Location = ResponseCacheLocation.None, NoStore = true)]
-        public ActionResult Comments()
-        {
-            return Json(_comments);
+            _logger = logger;
         }
 
         public IActionResult Index()
         {
-            return View(_comments);
+            return View();
         }
 
         public IActionResult Privacy()
