+++ WebApplicationMVC/Controllers/ReactJSController.cs
@@ -1,0 +1,64 @@
+using System.Diagnostics;
+using Microsoft.AspNetCore.Mvc;
+using WebApplicationMVC.Models;
+
+namespace WebApplicationMVC.Controllers
+{
+    public class ReactJSController : Controller
+    {
+        private static readonly IList<CommentModel> _comments;
+
+        static ReactJSController()
+        {
+            _comments = new List<CommentModel>
+            {
+                new CommentModel
+                {
+                    Id = 1,
+                    Author = "Daniel Lo Nigro",
+                    Text = "Hello ReactJS.NET World!"
+                },
+                new CommentModel
+                {
+                    Id = 2,
+                    Author = "Pete Hunt",
+                    Text = "This is one comment"
+                },
+                new CommentModel
+                {
+                    Id = 3,
+                    Author = "Jordan Walke",
+                    Text = "This is *another* comment"
+                },
+            };
+        }
+
+        [Route("comments/new")]
+        [HttpPost]
+        public ActionResult AddComment(CommentModel comment)
+        {
+            // Create a fake ID for this comment
+            comment.Id = _comments.Count + 1;
+            _comments.Add(comment);
+            return Content("Success :)");
+        }
+
+        [Route("comments")]
+        [ResponseCache(Location = ResponseCacheLocation.None, NoStore = true)]
+        public ActionResult Comments()
+        {
+            return Json(_comments);
+        }
+
+        public IActionResult ReactJS()
+        {
+            return View(_comments);
+        }
+
+        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
+        public IActionResult Error()
+        {
+            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
+        }
+    }
+}
