+++ WebApplicationMVC/Views/Home/Index.cshtml
@@ -1,23 +1,8 @@
-﻿@model IEnumerable<WebApplicationMVC.Models.CommentModel>
-@{
-    Layout = null;
+﻿@{
+    ViewData["Title"] = "Home Page";
 }
-<html>
-<head>
-    <title>Hello React</title>
-</head>
-<body>
-    @Html.React("CommentBox", new
-    {
-        initialData = Model,
-        url = Url.Action("Comments"),
-        submitUrl = Url.Action("AddComment"),
-        pollInterval = 2000
-    })
-    <script crossorigin src="https://cdnjs.cloudflare.com/ajax/libs/react/16.13.0/umd/react.development.js"></script>
-    <script crossorigin src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/16.13.0/umd/react-dom.development.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/remarkable/1.7.1/remarkable.min.js"></script>
-    <script src="@Url.Content("~/js/Tutorial.jsx")"></script>
-    @Html.ReactInitJavaScript()
-</body>
-</html>
+
+<div class="text-center">
+    <h1 class="display-4">Welcome</h1>
+    <p>Learn about <a href="https://learn.microsoft.com/aspnet/core">building Web apps with ASP.NET Core</a>.</p>
+</div>
