+++ WebApplicationMVC/Views/ReactJS/ReactJS.cshtml
@@ -1,0 +1,18 @@
+﻿@model IEnumerable<WebApplicationMVC.Models.CommentModel>
+@{
+    ViewBag.Title = "ReactJS";
+}
+@Html.React("CommentBox", new
+{
+    initialData = Model,
+    url = Url.Action("Comments"),
+    submitUrl = Url.Action("AddComment"),
+    pollInterval = 2000
+})
+@section Scripts {
+    <script crossorigin src="https://cdnjs.cloudflare.com/ajax/libs/react/16.13.0/umd/react.development.js"></script>
+    <script crossorigin src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/16.13.0/umd/react-dom.development.js"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/remarkable/1.7.1/remarkable.min.js"></script>
+    <script src="@Url.Content("~/js/Tutorial.jsx")"></script>
+    @Html.ReactInitJavaScript()
+}
