+++ WebApplicationMVC/Views/Shared/_Layout.cshtml
@@ -23,7 +23,10 @@
                             <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Index">Home</a>
                         </li>
                         <li class="nav-item">
-                            <a class="nav-link text-dark" asp-area="" asp-controller="Home" asp-action="Privacy">Privacy</a>
+                            <a class="nav-link text-dark" asp-area="" asp-controller="ReactJS" asp-action="ReactJS">ReactJS</a>
+                        </li>
+                        <li class="nav-item">
+                            <a class="nav-link text-dark" asp-area="" asp-controller="SignalR" asp-action="SignalR">SignalR</a>
                         </li>
                     </ul>
                     <partial name="_LoginPartial" />
