+++ WebApplicationMVC/Views/SignalR/SignalR.cshtml
@@ -1,5 +1,5 @@
 ﻿@{
-    ViewBag.Title = "Chat";
+    ViewBag.Title = "SignalR";
 }
 <h2>Chat</h2>
 <div class="container">
@@ -27,7 +27,7 @@
         </div>
     </div>
 </div>
-@section scripts {
+@section Scripts {
     <script src="~/js/signalr/dist/browser/signalr.js"></script>
     <script src="~/js/chat.js"></script>
 }
